+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * **New Resource:** `morpheus_max_cores_policy`
 * **New Resource:** `morpheus_max_hosts_policy`
 * **New Resource:** `morpheus_max_vms_policy`
+* **New Resource:** `morpheus_task_job`
 * **New Resource:** `morpheus_workflow_policy`
 
 ## 0.3.1 (September 23, 2021)